backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
beautifulsoup4==4.12.3
requests==2.32.3
google-generativeai==0.8.3
python-multipart==0.0.12
pydantic==2.9.2
lxml==5.3.0
python-dotenv==1.0.1
playwright==1.49.1
langgraph
